# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/texts/{text_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Text",
        "description": "Gets a text by the given id.",
        "operationId": "get_text",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextWithQuestionsAndStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Text",
        "description": "Deletes a text by the given id.",
        "operationId": "delete_text",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Text"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/current": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Current User",
        "description": "Gets the current user's information.",
        "operationId": "get_current_user",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/game/texts/{text_id}": {
      "get": {
        "tags": [
          "game"
        ],
        "summary": "Get Text By Id",
        "description": "Gets the text with the given id.",
        "operationId": "get_text_by_id",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextWithQuestions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/game/next-text": {
      "get": {
        "tags": [
          "game"
        ],
        "summary": "Get Next Text",
        "description": "Gets the next text that the user has not attempted before.",
        "operationId": "get_next_text",
        "parameters": [
          {
            "name": "difficulty",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Difficulty"
            }
          },
          {
            "name": "is_summary",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Is Summary"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Text"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/game/texts/{text_id}/next-questions": {
      "get": {
        "tags": [
          "game"
        ],
        "summary": "Get Next Questions",
        "description": "Gets next 10 questions that the user has not attempted before.",
        "operationId": "get_next_questions",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionMasked"
                  },
                  "title": "Response Get Next Questions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/game/texts/{text_id}/answers": {
      "post": {
        "tags": [
          "game"
        ],
        "summary": "Post Answers",
        "description": "Accepts the question answers and other statistics.\nReturns the results to the answers.",
        "operationId": "post_answers",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_answers"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Result"
                  },
                  "title": "Response Post Answers"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Answer": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "selected_option": {
            "type": "integer",
            "title": "Selected Option"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "selected_option"
        ],
        "title": "Answer"
      },
      "Body_post_answers": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/Answer"
            },
            "type": "array",
            "title": "Answers"
          },
          "average_wpm": {
            "type": "integer",
            "title": "Average Wpm"
          },
          "interval_wpms": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Interval Wpms"
          },
          "game_mode": {
            "type": "string",
            "title": "Game Mode"
          },
          "game_submode": {
            "type": "string",
            "title": "Game Submode"
          },
          "summary": {
            "type": "boolean",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "answers",
          "average_wpm",
          "interval_wpms",
          "game_mode",
          "game_submode",
          "summary"
        ],
        "title": "Body_post_answers"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Question": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "correct_option": {
            "type": "integer",
            "title": "Correct Option"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "content",
          "options",
          "correct_option",
          "id"
        ],
        "title": "Question"
      },
...
      "QuestionMasked": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content",
          "options"
        ],
        "title": "QuestionMasked"
      },
...
      "Result": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "correct_option": {
            "type": "integer",
            "title": "Correct Option"
          },
          "selected_option": {
            "type": "integer",
            "title": "Selected Option"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "correct",
          "correct_option",
          "selected_option"
        ],
        "title": "Result"
      },
...
      "Text": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fiction": {
            "type": "boolean",
            "title": "Fiction"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "summary",
          "source",
          "fiction",
          "difficulty",
          "word_count",
          "description",
          "author",
          "image_url",
          "id"
        ],
        "title": "Text"
      },
...
      "TextWithQuestions": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fiction": {
            "type": "boolean",
            "title": "Fiction"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "summary",
          "source",
          "fiction",
          "difficulty",
          "word_count",
          "description",
          "author",
          "image_url",
          "id",
          "questions"
        ],
        "title": "TextWithQuestions"
      },
      "TextWithQuestionsAndStatistics": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fiction": {
            "type": "boolean",
            "title": "Fiction"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions"
          },
          "min_wpm": {
            "type": "integer",
            "title": "Min Wpm"
          },
          "max_wpm": {
            "type": "integer",
            "title": "Max Wpm"
          },
          "average_wpm": {
            "type": "integer",
            "title": "Average Wpm"
          },
          "average_score": {
            "type": "integer",
            "title": "Average Score"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "summary",
          "source",
          "fiction",
          "difficulty",
          "word_count",
          "description",
          "author",
          "image_url",
          "id",
          "questions",
          "min_wpm",
          "max_wpm",
          "average_wpm",
          "average_score"
        ],
        "title": "TextWithQuestionsAndStatistics"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "role"
        ],
        "title": "User"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}